+++ src/settings/arguments.rs
@@ -2,12 +2,12 @@ use std::path::PathBuf;
 
 use clap::Parser;
 
-/// Play any game ROM with associated emulator in `"RetroArch"`.
+/// Play any game ROM with associated emulator in `RetroArch`.
 ///
 /// `enjoy` is a launcher to run games from `RetroArch` without using the GUI.  It is a wrapper
 /// around the `retroarch` commandline application.  The main functionality comes from a user
 /// configuration file with rules to associate file extensions and emulator cores.  When running
-/// **enjoy** only the path to a ROM file is required, either through arguments or the stdin.  Use
+/// `enjoy` only the path to a ROM file is required, either through arguments or the stdin.  Use
 /// option `-h` for compact summary or `--help` for longer description of options.
 /// <https://github.com/thingsiplay/enjoy/>
 ///
@@ -22,7 +22,8 @@ use clap::Parser;
 /// $ ls -1 $(readlink -f ~/roms/gb)/* | enjoy -xWn
 #[allow(clippy::struct_excessive_bools)]
 #[derive(Debug, Parser)]
-pub(crate) struct Opt {
+#[clap(version, author, after_help = "https://github.com/thingsiplay/enjoy")]
+pub struct Opt {
     /// Path to ROM file
     ///
     /// If multiple files are specified, then the first entry is picked when starting emulator.
@@ -32,7 +33,7 @@ pub(crate) struct Opt {
     ///
     /// Example: "~/roms/snes/Super Mario World (U) [\!].smc"
     #[clap(parse(from_os_str))]
-    pub(crate) games: Vec<PathBuf>,
+    pub games: Vec<PathBuf>,
 
     /// Bypass additional arguments to `retroarch`
     ///
@@ -43,7 +44,7 @@ pub(crate) struct Opt {
     ///
     /// Example: "-- --set-shader ''"
     #[clap(last = true)]
-    pub(crate) retroarch_arguments: Vec<String>,
+    pub retroarch_arguments: Vec<String>,
 
     /// Path to the user settings
     ///
@@ -60,7 +61,19 @@ pub(crate) struct Opt {
         display_order = 1,
         default_value = "~/.config/enjoy/default.ini"
     )]
-    pub(crate) config: PathBuf,
+    pub config: PathBuf,
+
+    /// Open user settings
+    ///
+    /// Opens the user config INI file with it's associated default application and exit.
+    #[clap(short = 'O', long, display_order = 1)]
+    pub open_config: bool,
+
+    /// Print path of user settings
+    ///
+    /// Prints path of the user config INI file to stdout and exit.
+    #[clap(short = 'o', long, display_order = 1)]
+    pub config_path: bool,
 
     /// Path or name of `RetroArch` command
     ///
@@ -73,9 +86,9 @@ pub(crate) struct Opt {
         long,
         parse(from_os_str),
         value_name = "APP",
-        display_order = 4
+        display_order = 7
     )]
-    pub(crate) retroarch: Option<PathBuf>,
+    pub retroarch: Option<PathBuf>,
 
     /// Path to `RetroArch` base configuration
     ///
@@ -90,9 +103,9 @@ pub(crate) struct Opt {
         long,
         parse(from_os_str),
         value_name = "FILE",
-        display_order = 4
+        display_order = 7
     )]
-    pub(crate) retroarch_config: Option<PathBuf>,
+    pub retroarch_config: Option<PathBuf>,
 
     /// Force specific libretro core by filename
     ///
@@ -109,10 +122,10 @@ pub(crate) struct Opt {
         long,
         parse(from_os_str),
         value_name = "FILE",
-        display_order = 3,
+        display_order = 5,
         conflicts_with = "core"
     )]
-    pub(crate) libretro: Option<PathBuf>,
+    pub libretro: Option<PathBuf>,
 
     /// Directory of libretro core files
     ///
@@ -126,9 +139,9 @@ pub(crate) struct Opt {
         long,
         parse(from_os_str),
         value_name = "DIR",
-        display_order = 3
+        display_order = 6
     )]
-    pub(crate) libretro_directory: Option<PathBuf>,
+    pub libretro_directory: Option<PathBuf>,
 
     /// Force specific libretro core by user defined alias
     ///
@@ -137,8 +150,8 @@ pub(crate) struct Opt {
     /// `[cores]` as `alias=libretro_path`.
     ///
     /// Example: "snes"
-    #[clap(short = 'C', long, value_name = "ALIAS", display_order = 3)]
-    pub(crate) core: Option<String>,
+    #[clap(short = 'C', long, value_name = "ALIAS", display_order = 4)]
+    pub core: Option<String>,
 
     /// Apply simple wildcard to filter list of games
     ///
@@ -146,52 +159,55 @@ pub(crate) struct Opt {
     /// functionality is limited and only the star `*` and questionmark `?` are supported.  The
     /// comparison is always case insensitive.  It will compare the base filename portion of the
     /// ROM path to the pattern, ignoring it's parent directory and filename extension.  At default
-    /// a star is added in front and end of pattern automatically when comparing.  This option is
-    /// useful if more than one game entry is given to the program.
+    /// a star is added in front and end of pattern automatically when comparing.  It is useful if
+    /// more than one game entry is given to the program.  This option can be specified multiple
+    /// times.  All of them have to match.
     ///
     /// Example: "mario*[\!]"
     #[clap(short = 'f', long, value_name = "PATTERN", display_order = 2)]
-    pub(crate) filter: Option<String>,
+    pub filter: Option<Vec<String>>,
+
+    /// Strict mode for filter
+    ///
+    /// Turns the option `--filter` to be more strict when comparing filenames.  It makes it case
+    /// sensitive and a word will match the beginning to end of filename, no longer are stars "*"
+    /// surrounding the search pattern added to match any part.
+    #[clap(short = 's', long, display_order = 2)]
+    pub strict: bool,
 
     /// Print selected game ROM
     ///
     /// Writes the full filepath of the selected game to stdout.
     #[clap(short = 'w', long, display_order = 1)]
-    pub(crate) which: bool,
+    pub which: bool,
 
     /// Print RetroArch commandline
     ///
     /// Writes full command with all arguments used to run RetroArch to stdout. Has higher priority
     /// than option --which.
     #[clap(short = 'W', long, display_order = 1)]
-    pub(crate) which_command: bool,
+    pub which_command: bool,
 
     /// Print all core names
     ///
-    /// Lists all core names on the left side of the user configuration under section "[cores]".
+    /// Lists all core names on the left side of the user configuration under section "\[cores\]".
     /// Will output matching cores to the libretro core that would be used with the game.  Without
     /// a game, all cores are listed.
-    #[clap(short = 'n', long, display_order = 2)]
-    pub(crate) list_cores: bool,
+    #[clap(short = 'n', long, display_order = 3)]
+    pub list_cores: bool,
 
     /// Force fullscreen mode
     ///
     /// Runs the emulator and `RetroArch` UI in fullscreen, regardless of any other setting.
-    #[clap(short = 'F', long, display_order = 2)]
-    pub(crate) fullscreen: bool,
+    #[clap(short = 'F', long, display_order = 3)]
+    pub fullscreen: bool,
 
     /// There Can Only Be One!
     ///
     /// Prevents running another `retroarch` process, if one is already active.  In this case the
     /// final command of the emulator will not execute.
-    #[clap(short = '1', long, display_order = 2)]
-    pub(crate) highlander: bool,
-
-    /// Show user settings
-    ///
-    /// Opens the user config INI file with it's associated default application.
-    #[clap(short = 'o', long, display_order = 3)]
-    pub(crate) open_config: bool,
+    #[clap(short = '1', long, display_order = 3)]
+    pub highlander: bool,
 
     /// Ignore user settings
     ///
@@ -201,23 +217,30 @@ pub(crate) struct Opt {
     #[clap(
         short = 'i',
         long,
-        display_order = 4,
+        display_order = 8,
         conflicts_with_all = &["config", "open-config", "core"]
     )]
-    pub(crate) noconfig: bool,
+    pub noconfig: bool,
 
     /// Do not run `RetroArch`
     ///
     /// The `retroarch` run command to play ROMs will not be executed.  Internally the process is
-    /// still simulated, up until to the point of running the emulator.
-    #[clap(short = 'x', long, display_order = 4)]
-    pub(crate) norun: bool,
+    /// still simulated, up until to the point of running the emulator.  If a game ROM is not
+    /// found, then the simulation will continue to allow execution of other options.
+    #[clap(short = 'x', long, display_order = 8)]
+    pub norun: bool,
 
     /// Dismiss reading from stdin
     ///
     /// Ignores the `stdin` and do not test or read any data from it.  Normally the program will
     /// look and read all lines from `stdin` as additional game entries.  This option will disable
     /// that.
-    #[clap(short = 'z', long, display_order = 4)]
-    pub(crate) nostdin: bool,
+    #[clap(short = 'z', long, display_order = 8)]
+    pub nostdin: bool,
+
+    /// Print version information
+    ///
+    /// Print the version number of this app and exit
+    #[clap(short = 'v', long, display_order = 9)]
+    pub version: bool,
 }