+++ src/settings/retroarch.rs
@@ -1,16 +1,16 @@
 use crate::settings::file;
 
-use std::collections::HashMap;
 use std::collections::HashSet;
 use std::error::Error;
 use std::path::PathBuf;
 use std::process::Command;
 
 use configparser::ini;
+use indexmap::map::IndexMap;
 
 /// Check if a process is running.  If `print_pid` is `true`, then print the pid of found process
 /// to stdout.
-pub(crate) fn is_running(process_name: &str, print_pid: bool) -> bool {
+pub fn is_running(process_name: &str, print_pid: bool) -> bool {
     let mut cmdline: Command = Command::new(String::from("pidof"));
 
     // return one PID only
@@ -34,7 +34,7 @@ pub(crate) fn is_running(process_name: &str, print_pid: bool) -> bool {
 ///     2. `~/.config/retroarch/retroarch.cfg`
 ///     3. `~/.retroarch.cfg`
 /// ... in that order.
-pub(crate) fn search_default_config() -> Option<PathBuf> {
+pub fn search_default_config() -> Option<PathBuf> {
     let mut fullpath: PathBuf;
 
     if let Ok(path) =
@@ -62,14 +62,14 @@ pub(crate) fn search_default_config() -> Option<PathBuf> {
     None
 }
 
-/// Parses a `RetroArch` configuration file and returns a `HashMap` from it.  The format is like
+/// Parses a `RetroArch` configuration file and returns a `IndexMap` from it.  The format is like
 /// a regular INI file without sections.  The set `lookup_keys` contains all key names to look
 /// for in the file and extract only those key and value pairs as strings.  The surrounding
 /// double quotes are removed from the value.
-pub(crate) fn parse_retroarch_config(
+pub fn parse_retroarch_config(
     path: &Option<PathBuf>,
     lookup_keys: &HashSet<String>,
-) -> Result<HashMap<String, String>, Box<dyn Error>> {
+) -> Result<IndexMap<String, String>, Box<dyn Error>> {
     let mut ini = ini::Ini::new_cs();
 
     match ini.load(
@@ -85,12 +85,12 @@ pub(crate) fn parse_retroarch_config(
 }
 
 // Searches all `lookup_keys` in `default` section of an INI structure and returns a regular
-// HashMap of it.  Empty strings or missing keys are excluded.
+// IndexMap of it.  Empty strings or missing keys are excluded.
 fn extract_default_inikeys(
-    ini: &HashMap<String, HashMap<String, Option<String>>>,
+    ini: &IndexMap<String, IndexMap<String, Option<String>>>,
     lookup_keys: &HashSet<String>,
-) -> HashMap<String, String> {
-    let mut found_keys: HashMap<String, String> = HashMap::new();
+) -> IndexMap<String, String> {
+    let mut found_keys: IndexMap<String, String> = IndexMap::new();
 
     for (key, value) in ini
         .get("default")
@@ -108,7 +108,7 @@ fn extract_default_inikeys(
 /// Combine the `libretro-directory` and `libretro` core file to a fullpath.  Add a string to
 /// the end of the filename, if it does not end like that.  This includes the file extension
 /// and end of the filename part.  In example the common "_libretro.so" could be added.
-pub(crate) fn libretro_fullpath(
+pub fn libretro_fullpath(
     directory: Option<PathBuf>,
     libretro: Option<PathBuf>,
     endswith: &str,
@@ -127,10 +127,10 @@ pub(crate) fn libretro_fullpath(
 #[cfg(test)]
 mod tests {
 
-    use std::collections::HashMap;
     use std::collections::HashSet;
 
     use configparser::ini;
+    use indexmap::map::IndexMap;
 
     // Untested:
     //  - search_default_config()
@@ -148,7 +148,7 @@ mod tests {
 
     #[test]
     fn extract_default_inikeys_single() {
-        let inidata: HashMap<String, HashMap<String, Option<String>>>;
+        let inidata: IndexMap<String, IndexMap<String, Option<String>>>;
         inidata = ini::Ini::new()
             .read(String::from(
                 "
@@ -174,7 +174,7 @@ mod tests {
 
     #[test]
     fn extract_default_inikeys_multiple() {
-        let inidata: HashMap<String, HashMap<String, Option<String>>>;
+        let inidata: IndexMap<String, IndexMap<String, Option<String>>>;
         inidata = ini::Ini::new()
             .read(String::from(
                 "